Fix address discovery panic when canceling sync
Address discovery thread crashes due to nil activeSyncData which happens
because the timer ticked during the sync cancellation process. The fix
was implemented by using a mutex to hold the access to syncData and also
check if activeSyncData is nil before access it's data during address
discovery.

This commit also
- adds WalletName property to LibWallet struct and set default wallet
  name to "wallet-#".
- fixes a bug in fetch header progress notification being sent before
  inserting the blocks into the wallet which caused the wallet's best
  block to be lesser than the total fetched headers.
- adds waiting boolean property to syncData which is set to true by
  default and then set to true as soon as the wallets joins the fetch
  header process.
- adds a sync check before creating wallet and would error if sync
  is active. This would help adding a newly created wallet to the syncer.